+++ tests/entrypoints/openai/responses/test_errors.py
