+++ tests/entrypoints/openai/responses/test_mcp_tools.py
@@ -9,7 +9,7 @@ from openai_harmony import ToolDescription, ToolNamespaceConfig
 
 from vllm.entrypoints.tool_server import MCPToolServer
 
-from ...utils import RemoteOpenAIServer
+from ....utils import RemoteOpenAIServer
 
 MODEL_NAME = "openai/gpt-oss-20b"
