+++ tests/entrypoints/openai/responses/test_parsable_context.py
@@ -8,7 +8,7 @@ import pytest
 import pytest_asyncio
 from openai import OpenAI
 
-from ...utils import RemoteOpenAIServer
+from ....utils import RemoteOpenAIServer
 
 MODEL_NAME = "Qwen/Qwen3-8B"
